Add dconf dump to repo
There is no clear cut way as to what the best way to add the dconf
database is. Certainly the easiest would be to directly commit the
binary file, but that would not make for a readable history. Addding
the dump takes a bit more manual work but it is much cleaner.

As the settings accumulate over time it might make more sense to
break the dump into smaller segments.